code improvements after review
